Added some alerts

+++ frontend/src/app/all-properties/all-properties.component.ts
@@ -25,8 +25,14 @@ export class AllPropertiesComponent implements OnInit {
   }
 
   delete(id: number) {
-    this.pService.deleteProperty(id).subscribe((data) => {
+    if(confirm("Are you sure you want to delete?")){
+      this.pService.deleteProperty(id).subscribe((data) => {
+        this.getAll();
+      });
+    }
+    else{
       this.getAll();
-    });
+    }
+    
   }
 }
